requirements.txt

aiogram>=3.0
aiosqlite>=0.19
python-dotenv>=1.0
aiohttp>=3.9
